+++ WinUIGallery/Helper/WindowHelper.cs
@@ -19,7 +19,6 @@ using System.Threading.Tasks;
 using Windows.Storage;
 using WinRT;
 using WinRT.Interop;
-using static WinUIGallery.Win32;
 
 namespace WinUIGallery.Helper
 {
@@ -101,28 +100,6 @@ namespace WinUIGallery.Helper
                 localFolder = ApplicationData.Current.LocalFolder;
             }
             return localFolder;
-        }
-
-        public static void SetTitleBarBackdrop(this Window window, SystemBackdrop backdrop)
-        {
-            var backdropType = DwmSystemBackdropType.DWMSBT_AUTO;
-
-            if (backdrop is MicaBackdrop micaBackdrop)
-                backdropType = micaBackdrop.Kind == MicaKind.Base ? DwmSystemBackdropType.DWMSBT_MAINWINDOW : DwmSystemBackdropType.DWMSBT_TABBEDWINDOW;
-            else if (backdrop is DesktopAcrylicBackdrop)
-                backdropType = DwmSystemBackdropType.DWMSBT_TRANSIENTWINDOW;
-            else if (backdrop is null)
-                backdropType = DwmSystemBackdropType.DWMSBT_NONE;
-
-            var backdropTypeInt = ((int)backdropType);
-            int hResult = DwmSetWindowAttribute(WindowNative.GetWindowHandle(window), DwmWindowAttribute.DWMWA_SYSTEMBACKDROP_TYPE, ref backdropTypeInt, sizeof(DwmSystemBackdropType));
-            ExceptionHelpers.ThrowExceptionForHR(hResult);
-        }
-
-        public static void SetTitleBarTheme(this Window window)
-        {
-            int darkMode = ((FrameworkElement)window.Content).ActualTheme == ElementTheme.Dark ? 1 : 0;
-            DwmSetWindowAttribute(WindowNative.GetWindowHandle(window), DwmWindowAttribute.DWMWA_USE_IMMERSIVE_DARK_MODE, ref darkMode, sizeof(int));
         }
     }
 }
